+++ lib/rom-rails.rb
@@ -1,8 +1,3 @@
-module ROM
-  module Rails
-  end
-end
-
 require 'rom'
 require 'rom/rails/version'
 require 'rom/rails/railtie'
